+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s7/s6/s8/s_0768_test.kt
@@ -1,0 +1,16 @@
+package s0.s7.s6.s8
+
+import utils.T
+import kotlin.test.Test
+import s0.s7.s6.s8.s_0768_00.Solution as s_0768_00
+
+class s_0768_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0768_00::class,
+        )
+        t.run()
+    }
+}