+++ workspace/solutions/rust/src/s0/s7/mod.rs
@@ -1,2 +1,3 @@
 pub mod s1;
 pub mod s3;
+pub mod s6;