+++ workspace/solutions/rust/src/s0/s7/s6/s8/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0768_00;
+pub mod s_0768_test;
+
+pub type TSolution = fn(Vec<i32>) -> i32;